重构部分时间检测处理逻辑
